airtravel/air_travel_back/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
websockets
requests
authlib
google-auth
google-auth-oauthlib
fastapi-mail

# Translation services
azure-ai-formrecognizer
azure-ai-translation-text
azure-cognitiveservices-speech
azure-ai-documentintelligence
Pillow
numpy
pydub

# AI Models
torch
torchvision
clip-by-openai
faiss-cpu
ultralytics
openai
azure-cognitiveservices-speech